requirements.txt
alabaster==0.7.9
appdirs==1.4.0
Babel==2.3.4
docutils==0.13.1
graphviz==0.5.2
h5py==2.6.0
imagesize==0.7.1
Jinja2==2.9.5
Keras==1.2.2
MarkupSafe==0.23
nose==1.3.7
nose-parameterized==0.5.0
numpy==1.12.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
protobuf==3.2.0
pydot-ng==1.0.0
Pygments==2.2.0
pyparsing==2.1.10
pytz==2016.10
PyYAML==3.12
requests==2.13.0
scipy==0.18.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.5.2
tensorflow-gpu==0.12.1
Theano==0.8.2
ujson==1.35
